requirements.txt
# Core dependencies
requests>=2.28.1
feedparser>=6.0.10
newspaper3k>=0.2.8
python-dotenv>=0.21.0
SQLAlchemy>=2.0.0
APScheduler>=3.10.1

# AI and ML dependencies
google-generativeai>=0.3.0
scikit-learn>=1.3.0
numpy>=1.24.0

# HTML parsing
beautifulsoup4>=4.11.1
lxml>=4.9.2

# Data handling
tqdm>=4.64.1 